build: factor I/O functions from central ddbridge.h header into ddbridge-io.c/h
This tidies up ddbridge.h from all functions (especially those carrying
comparably complex code) and moves them into ddbridge-io.c/h.

Added benefit: ddbridge and octonet build is now free from warnings.
