+++ ddbridge/Kbuild
@@ -1,7 +1,7 @@
 EXTRA_CFLAGS += -DCONFIG_DVB_CXD2843 -DCONFIG_DVB_LNBP21 -DCONFIG_DVB_STV090x -DCONFIG_DVB_STV6110x -DCONFIG_DVB_DRXK -DCONFIG_DVB_STV0910 -DCONFIG_DVB_STV6111 -DCONFIG_DVB_LNBH25 -DCONFIG_DVB_MXL5XX
 
-ddbridge-objs := ddbridge-main.o ddbridge-core.o ddbridge-hw.o ddbridge-i2c.o ddbridge-modulator.o ddbridge-ns.o
-octonet-objs := octonet-main.o ddbridge-core.o ddbridge-hw.o ddbridge-i2c.o ddbridge-modulator.o ddbridge-ns.o
+ddbridge-objs := ddbridge-main.o ddbridge-core.o ddbridge-hw.o ddbridge-i2c.o ddbridge-io.o ddbridge-modulator.o ddbridge-ns.o
+octonet-objs := octonet-main.o ddbridge-core.o ddbridge-hw.o ddbridge-i2c.o ddbridge-io.o ddbridge-modulator.o ddbridge-ns.o
 
 obj-$(CONFIG_DVB_DDBRIDGE) += ddbridge.o
 obj-$(CONFIG_DVB_OCTONET) += octonet.o
