+++ ddbridge/ddbridge-i2c.c
@@ -39,6 +39,7 @@
 
 #include "ddbridge.h"
 #include "ddbridge-regs.h"
+#include "ddbridge-io.h"
 
 /******************************************************************************/
