+++ ddbridge/ddbridge-modulator.c
@@ -25,6 +25,7 @@
 
 #include "ddbridge.h"
 #include "ddbridge-regs.h"
+#include "ddbridge-io.h"
 
 #include <linux/dvb/mod.h>
